allow validators to transform result
