+++ client/package.json
@@ -80,8 +80,8 @@
     "@types/react-dom": "18.0.8",
     "@types/tinycolor2": "1.4.3",
     "@types/toposort": "2.0.3",
-    "@typescript-eslint/eslint-plugin": "5.42.0",
-    "@typescript-eslint/parser": "5.42.0",
+    "@typescript-eslint/eslint-plugin": "5.44.0",
+    "@typescript-eslint/parser": "5.44.0",
     "@vitejs/plugin-react": "2.2.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-typescript": "17.0.0",
